backend/requirements.txt
# Web Framework
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0

# Image Processing
pillow>=10.2.0,<11.0.0

# AWS & Database
boto3>=1.34.0,<2.0.0
supabase>=2.3.0,<3.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0

# Validation & Utils
python-multipart>=0.0.9,<1.0.0
pydantic>=2.5.0,<3.0.0
httpx>=0.27.0,<1.0.0